+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ unreleased
 * Removed useless console_script entry point.
 * Changed loader detection to only consider the final extension in a filename.
 * Support and test egg specifications with no named entry point (because it's 'main').
+* Support basic interpolation (``here`` and ``__file__``) in the built-in INI loader.
 
 0.2.0 (2015-06-14)
 -----------------------------------------